move get_static_path from conftest to tests.unit
