+++ tests/unit/conftest.py
@@ -18,6 +18,7 @@ from pycontrails.core import cache, met, met_var
 from pycontrails.core.aircraft_performance import AircraftPerformance, AircraftPerformanceGrid
 from pycontrails.datalib.ecmwf import ERA5
 from tests import BADA3_PATH, BADA4_PATH, BADA_AVAILABLE
+from tests.unit import get_static_path
 
 # find default cache dir for testing
 DISK_CACHE_DIR = cache._get_user_cache_dir()
@@ -38,22 +39,6 @@ def pytest_addoption(parser: pytest.Parser) -> None:
 def regenerate_results(request: pytest.FixtureRequest) -> Any:
     """Regenerate static test results in tests that use them."""
     return request.config.getoption("--regenerate-results")
-
-
-def get_static_path(filename: str | pathlib.Path) -> pathlib.Path:
-    """Return a path to file in ``/tests/static/`` directory.
-
-    Parameters
-    ----------
-    filename : str | pathlib.Path
-        Filename to prefix
-
-    Returns
-    -------
-    pathlib.Path
-    """
-    parent = pathlib.Path(__file__).parent
-    return parent / "static" / filename
 
 
 @pytest.fixture(scope="session")
